added language parameter for user in backend simplified token generation
