+++ backend/modules/database_init.ts
@@ -21,6 +21,7 @@ export const database_init = async () => {
                 password TEXT NOT NULL,
                 createdAt DATETIME DEFAULT CURRENT_TIMESTAMP,
                 verified BOOLEAN NOT NULL DEFAULT 0,
+                language TEXT NOT NULL DEFAULT 'fr',
                 publishToLeaderboard BOOLEAN DEFAULT NULL
             );
             CREATE TABLE IF NOT EXISTS tokens (
